Use size_t to match strlen return
(cherry picked from commit 3bb53712c3a1698d992bbd48bc4c1fea1f9324c2)

+++ include/OTUTOR.h
@@ -59,7 +59,7 @@ struct TutorTextBlock
 	enum { BUTTON_CODE_LEN=8 };
 
 	const char* text_ptr;          // offset of the help text in the text buffer
-	short text_len;          // length of the help text
+	size_t text_len;               // length of the help text
 
 	char	button_code[BUTTON_CODE_LEN+1];
 };
